requirements.txt
django
Pillow
django-crispy-forms
crispy-bootstrap5
django-filter